pyproject.toml
[project]
name = "password-vault-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asyncpg (>=0.30.0,<1.0)",
    "fastapi>=0.115.11",
    "pydantic>=2.10.6",
    "uvicorn>=0.34.0",
    "redis (>=5.2.1,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<1.0.0)",
    "PyJWT (>=2.10.1,<3.0.0)",
    "pytz (>=2025.1,<2026.0)"
]

[dependency-groups]
docker = [
    "maturin>=1.8.3"
]
dev = [
    "maturin>=1.8.3",
    "ruff>=0.10.0",
    "pre-commit (>=4.2,<5.0)",
    "pyright>=1.1.398",
    "coverage (>=7.8.0,<8.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["app"]

[tool.pyright]
include = ["app"]

[tool.coverage.run]
omit = ["*/site-packages/*", "tests/*"]
command_line = "-m pytest tests"
source = ["./app"]

[tool.coverage.report]
include = ["app/*"]


[tool.ruff]
line-length = 120
target-version = "py313"
lint.select = [
    "E",
    "F",
    "I",
    "A",
    "B",
    "ANN",
    "FAST",
    "ASYNC",
    "ARG",
    "S",
    "BLE",
    "COM",
    "C4",
    "T20",
    "Q",
    "RET",
    "SIM",
    "TC",
    "PTH",
    "N",
    "PERF",
    "W",
    "RUF"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "app"
module-name = "lwe"
